Add GuessParams([]io.Reader)
Given a bunch of Readers, read the first fragment header from each;
check that they all have the same backend, fragment size, and original
data size; and return an ErasureCodeParams instance with an appropriate
backend name and an upper-bound for the number of parity fragments.

Currently, the caller needs to know the appropriate number of data
fragments -- in the future we might be able to guess this from the ratio
of original data size to fragment size.