Prevent early release of memory by GC
Adds a new test on decode: run multiple goroutines in parallel to force
the GC to wake up and possibly reallocate in-use memory

The apparent issue with GC is that allocated blocks are no longer in use
from go point of view (but actually are referenced by C layer).

Adding a runtime.KeepAlive(block) prevents the GC to release/reallocate
the memory.

Impacts Reconstruct & Decode